# perf(streams): make `toBlob()` 20-30% faster