It's faster just to use `Response.blob()`. This also makes the argument more stricter by only allowing `Uint8Array`s, which the previous implementation implied. Inspired by talks in #3894.

Benchmark source code:
```ts
import { toBlob } from "https://deno.land/std@0.208.0/streams/mod.ts";

Deno.bench("toBlob()", { group: "blob" }, async () => {
  const stream = ReadableStream.from([
    new Uint8Array([1, 2, 3, 4, 5]),
    new Uint8Array([6, 7]),
    new Uint8Array([8, 9]),
  ]);

  await toBlob(stream);
});

Deno.bench("Response.blob()", { group: "blob", baseline: true }, async () => {
  const stream = ReadableStream.from([
    new Uint8Array([1, 2, 3, 4, 5]),
    new Uint8Array([6, 7]),
    new Uint8Array([8, 9]),
  ]);

  await new Response(stream).blob();
});
```

Benchmark results:
```
cpu: Apple M2
runtime: deno 1.38.5 (aarch64-apple-darwin)

file:///Users/asher/GitHub/deno_std/bench.ts
benchmark            time (avg)        iter/s             (min … max)       p75       p99      p995
--------------------------------------------------------------------- -----------------------------


toBlob()              5.89 µs/iter     169,894.7      (4.25 µs … 4.6 ms)   5.33 µs  10.08 µs  12.75 µs
Response.blob()       4.68 µs/iter     213,504.9     (4.37 µs … 6.61 µs)   4.55 µs   6.61 µs   6.61 µs

summary
  Response.blob()
   1.26x faster than toBlob()
```