@kt3k 2023-12-12
> This also makes the argument more stricter by only allowing Uint8Arrays, which the previous implementation implied.

Does the previous implementation imply narrower typing? It looks supporting string inputs