@iuioiua 2023-12-12
The function does support string streams. However, it's unclear whether this is deliberate. Reasons being:
1. The argument accepts `ReadableStream` without a defined type variable.
2. The JSDoc says, `Works the same as {@linkcode Response.blob}`, which strictly only accepts `ReadableStream<Uint8Array>`.
3. The JSDocs also says it supports string streams.
4. There are no tests for string stream arguments.

We need to make a decision on whether it should support `ReadableStream<string | Uint8Array>` or `ReadableStream<Uint8Array>`.

I think supporting only `ReadableStream<Uint8Array>` and having the simple implementation of using `Response.blob()` under the hood is the way to go. Doing so will make the design and expected behaviour of the function clear and eliminate future confusion. Tweaking code that uses this function is easily done too.

Again, I think that these `toJson()`, `toText()`, etc. functions are redundant. But having them perfectly mirror `Response` methods gives them a clear definition and simple implementation.