@kt3k 2023-12-14
Now I think this is a very good idea as this doesn't load everything on-memory.

Streaming file upload with `FormData` can be done like:

```js
import { toBlob } from "https://deno.land/std@0.209.0/streams/mod.ts";

const file = await Deno.open("huge-file");
const form = new FormData();
form.set("file", toBlob(file.readable));
await fetch("url", { method: "POST", body: form });
```

With the current implementation, the above loads and stores the entire file contents in `chunks` in `toBlob` call, which is very inefficient when the file is large.